// src/lib/types/db/purple/purpleParty.ts
export interface PurplePartyDbModel {
	RunID: string;
	P1PlayerID: string;
	P2PlayerID: string;
	P3PlayerID: string;
	P4PlayerID: string;
	P1RunCharacter: string;
	P2RunCharacter: string;
	P3RunCharacter: string;
	P4RunCharacter: string;
	Patch: string;
	Region: string;
	Rank: string;
	Time: string;
	P1MainClass: string;
	P2MainClass: string;
	P3MainClass: string;
	P4MainClass: string;
	P1SubClass: string;
	P2SubClass: string;
	P3SubClass: string;
	P4SubClass: string;
	PartySize: string;
	P1Link: string;
	P2Link: string;
	P3Link: string;
	P4Link: string;
	Notes: string;
	SubmissionTime: string;
	SubmitterID: string;
	ServerID: string;
	P1PlayerName: string;
	P1PlayerCName: string;
	P1PlayerNameType: string;
	P1PlayerNameColor1: string;
	P1PlayerNameColor2: string;
	P1PlayerServer: string;
	P1PlayerPrefN: string;
	P2PlayerName: string;
	P2PlayerCName: string;
	P2PlayerNameType: string;
	P2PlayerNameColor1: string;
	P2PlayerNameColor2: string;
	P2PlayerServer: string;
	P2PlayerPrefN: string;
	P3PlayerName: string;
	P3PlayerCName: string;
	P3PlayerNameType: string;
	P3PlayerNameColor1: string;
	P3PlayerNameColor2: string;
	P3PlayerServer: string;
	P3PlayerPrefN: string;
	P4PlayerName: string;
	P4PlayerCName: string;
	P4PlayerNameType: string;
	P4PlayerNameColor1: string;
	P4PlayerNameColor2: string;
	P4PlayerServer: string;
	P4PlayerPrefN: string;
	SubmitterName: string;
	SubmitterCName: string;
	SubmitterNameType: string;
	SubmitterNameColor1: string;
	SubmitterNameColor2: string;
	SubmitterPrefN: string;
}